requirements.txt
wget
transformers
sentencepiece
wrapt
omegaconf
pytorch-lightning===1.5.6
hydra-core
fuzzywuzzy
uvicorn
fastapi
joblib
redis==4.5.5
redis-streams==0.2.0
python-dateutil
pyjwt[crypto]
accelerate
python-Levenshtein
